fix issues

+++ database/migrations/2026_02_01_201827_add_type_doctor_id_to_notifications_table.php
@@ -11,10 +11,12 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('notifications', function (Blueprint $table) {
-            $table->unsignedBigInteger('type_doctor_id')->nullable()->after('doctor_id');
-            $table->foreign('type_doctor_id')->references('id')->on('users')->onDelete('set null');
-        });
+        if (!Schema::hasColumn('notifications', 'type_doctor_id')) {
+            Schema::table('notifications', function (Blueprint $table) {
+                $table->unsignedBigInteger('type_doctor_id')->nullable()->after('doctor_id');
+                $table->foreign('type_doctor_id')->references('id')->on('users')->onDelete('set null');
+            });
+        }
     }
 
     /**
